+++ saleor/graphql/product/bulk_mutations/products.py
@@ -3,9 +3,13 @@ from collections import defaultdict
 import graphene
 from django.core.exceptions import ValidationError
 from django.db import transaction
+from django.db.models import Q
 from graphene.types import InputObjectType
 
+from ....attribute import AttributeInputType
+from ....attribute import models as attribute_models
 from ....core.permissions import ProductPermissions, ProductTypePermissions
+from ....core.tracing import traced_atomic_transaction
 from ....order import OrderStatus
 from ....order import models as order_models
 from ....order.tasks import recalculate_orders_task
@@ -37,7 +41,7 @@ from ..mutations.products import (
     ProductVariantInput,
     StockInput,
 )
-from ..types import Product, ProductVariant
+from ..types import Product, ProductType, ProductVariant
 from ..utils import create_stocks, get_used_variants_attribute_values
 
 
@@ -99,7 +103,7 @@ class ProductBulkDelete(ModelBulkDeleteMutation):
         error_type_field = "product_errors"
 
     @classmethod
-    @transaction.atomic
+    @traced_atomic_transaction()
     def perform_mutation(cls, _root, info, ids, **data):
         _, pks = resolve_global_ids_to_primary_keys(ids, Product)
         product_to_variant = list(
@@ -115,6 +119,7 @@ class ProductBulkDelete(ModelBulkDeleteMutation):
         ).values("pk", "order_id")
         line_pks = {line["pk"] for line in lines_id_and_orders_id}
         orders_id = {line["order_id"] for line in lines_id_and_orders_id}
+        cls.delete_assigned_attribute_values(pks)
         response = super().perform_mutation(
             _root,
             info,
@@ -129,6 +134,13 @@ class ProductBulkDelete(ModelBulkDeleteMutation):
             recalculate_orders_task.delay(list(orders_id))
 
         return response
+
+    @staticmethod
+    def delete_assigned_attribute_values(instance_pks):
+        attribute_models.AttributeValue.objects.filter(
+            productassignments__product_id__in=instance_pks,
+            attribute__input_type__in=AttributeInputType.TYPES_WITH_UNIQUE_VALUES,
+        ).delete()
 
     @classmethod
     def bulk_action(cls, info, queryset, product_to_variant):
@@ -437,7 +449,7 @@ class ProductVariantBulkCreate(BaseMutation):
         )
 
     @classmethod
-    @transaction.atomic
+    @traced_atomic_transaction()
     def save_variants(cls, info, instances, product, cleaned_inputs):
         assert len(instances) == len(
             cleaned_inputs
@@ -463,7 +475,7 @@ class ProductVariantBulkCreate(BaseMutation):
         create_stocks(variant, stocks, warehouses)
 
     @classmethod
-    @transaction.atomic
+    @traced_atomic_transaction()
     def perform_mutation(cls, root, info, **data):
         product = cls.get_node_or_error(info, data["product_id"], models.Product)
         errors = defaultdict(list)
@@ -509,7 +521,7 @@ class ProductVariantBulkDelete(ModelBulkDeleteMutation):
         error_type_field = "product_errors"
 
     @classmethod
-    @transaction.atomic
+    @traced_atomic_transaction()
     def perform_mutation(cls, _root, info, ids, **data):
         _, pks = resolve_global_ids_to_primary_keys(ids, ProductVariant)
         # get draft order lines for variants
@@ -533,6 +545,7 @@ class ProductVariantBulkDelete(ModelBulkDeleteMutation):
             )
         )
 
+        cls.delete_assigned_attribute_values(pks)
         response = super().perform_mutation(_root, info, ids, **data)
 
         transaction.on_commit(
@@ -556,6 +569,13 @@ class ProductVariantBulkDelete(ModelBulkDeleteMutation):
             product.save(update_fields=["default_variant"])
 
         return response
+
+    @staticmethod
+    def delete_assigned_attribute_values(instance_pks):
+        attribute_models.AttributeValue.objects.filter(
+            variantassignments__variant_id__in=instance_pks,
+            attribute__input_type__in=AttributeInputType.TYPES_WITH_UNIQUE_VALUES,
+        ).delete()
 
 
 class ProductVariantStocksCreate(BaseMutation):
@@ -667,7 +687,7 @@ class ProductVariantStocksUpdate(ProductVariantStocksCreate):
         return cls(product_variant=variant)
 
     @classmethod
-    @transaction.atomic
+    @traced_atomic_transaction()
     def update_or_create_variant_stocks(cls, variant, stocks_data, warehouses):
         stocks = []
         for stock_data, warehouse in zip(stocks_data, warehouses):
@@ -729,6 +749,23 @@ class ProductTypeBulkDelete(ModelBulkDeleteMutation):
         permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
         error_type_class = ProductError
         error_type_field = "product_errors"
+
+    @classmethod
+    @traced_atomic_transaction()
+    def perform_mutation(cls, _root, info, ids, **data):
+        _, pks = resolve_global_ids_to_primary_keys(ids, ProductType)
+        cls.delete_assigned_attribute_values(pks)
+        return super().perform_mutation(_root, info, ids, **data)
+
+    @staticmethod
+    def delete_assigned_attribute_values(instance_pks):
+        attribute_models.AttributeValue.objects.filter(
+            Q(attribute__input_type__in=AttributeInputType.TYPES_WITH_UNIQUE_VALUES)
+            & (
+                Q(productassignments__assignment__product_type_id__in=instance_pks)
+                | Q(variantassignments__assignment__product_type_id__in=instance_pks)
+            )
+        ).delete()
 
 
 class ProductMediaBulkDelete(ModelBulkDeleteMutation):