+++ saleor/product/utils/__init__.py
@@ -1,8 +1,7 @@
 from typing import TYPE_CHECKING, Dict, Iterable, List, Union
 
-from django.db import transaction
-
 from ...core.taxes import TaxedMoney, zero_taxed_money
+from ...core.tracing import traced_atomic_transaction
 from ..models import Product, ProductChannelListing
 from ..tasks import update_products_discounted_prices_task
 
@@ -32,7 +31,7 @@ def calculate_revenue_for_variant(
     return revenue
 
 
-@transaction.atomic
+@traced_atomic_transaction()
 def delete_categories(categories_ids: List[str], manager):
     """Delete categories and perform all necessary actions.